requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
aio-pika==9.0.4
anyio==3.6.2
astroid==2.13.3
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
discord.py==2.1.0
fastapi==0.89.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
isort==5.11.4
lazy-object-proxy==1.9.0
loguru==0.6.0
mccabe==0.7.0
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
packaging==23.0
pika==1.3.1
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
psycopg2==2.9.5
pydantic==1.10.4
pylint==2.15.10
pypika-tortoise==0.1.6
pytest==7.2.1
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.1
sqlalchemy2-stubs==0.0.2a32
starlette==0.22.0
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
yarl==1.8.2
